# shared/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RAG Corp API",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/ingest/text": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest Text",
        "operationId": "ingest_text_v1_ingest_text_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestTextReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestTextRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ingest/batch": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "Ingest Batch",
        "description": "Ingest multiple documents in a single request.\n\nProcesses up to 10 documents sequentially.\nReturns results for all successfully ingested documents.",
        "operationId": "ingest_batch_v1_ingest_batch_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBatchReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/query": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Query",
        "operationId": "query_v1_query_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/ask": {
      "post": {
        "tags": [
          "query"
        ],
        "summary": "Ask",
        "description": "R: RAG endpoint using Clean Architecture (Use Case pattern).\n\nThis endpoint demonstrates the architecture improvement:\n- Business logic in use case (testable, framework-independent)\n- Dependency injection via FastAPI Depends\n- Separation of concerns (HTTP ↔ Business Logic)\n\nUses the same query contract as /query with a generation step.",
        "operationId": "ask_v1_ask_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest/text": {
      "post": {
        "tags": [
          "v1",
          "ingest"
        ],
        "summary": "Ingest Text",
        "operationId": "ingest_text_api_v1_ingest_text_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestTextReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestTextRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ingest/batch": {
      "post": {
        "tags": [
          "v1",
          "ingest"
        ],
        "summary": "Ingest Batch",
        "description": "Ingest multiple documents in a single request.\n\nProcesses up to 10 documents sequentially.\nReturns results for all successfully ingested documents.",
        "operationId": "ingest_batch_api_v1_ingest_batch_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestBatchReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/query": {
      "post": {
        "tags": [
          "v1",
          "query"
        ],
        "summary": "Query",
        "operationId": "query_api_v1_query_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ask": {
      "post": {
        "tags": [
          "v1",
          "query"
        ],
        "summary": "Ask",
        "description": "R: RAG endpoint using Clean Architecture (Use Case pattern).\n\nThis endpoint demonstrates the architecture improvement:\n- Business logic in use case (testable, framework-independent)\n- Dependency injection via FastAPI Depends\n- Separation of concerns (HTTP ↔ Business Logic)\n\nUses the same query contract as /query with a generation step.",
        "operationId": "ask_api_v1_ask_post",
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Api-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AskRes"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/healthz": {
      "get": {
        "summary": "Healthz",
        "description": "R: Enhanced health check that verifies system dependencies.\n\nArgs:\n    full: If True, also check Google API connectivity (slower)\n\nReturns:\n    ok: True if all checked systems operational\n    db: \"connected\" or \"disconnected\"\n    google: \"available\", \"unavailable\", or \"disabled\" (only with full=true)\n    request_id: Correlation ID for this request",
        "operationId": "healthz_healthz_get",
        "parameters": [
          {
            "name": "full",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Full"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Metrics",
        "description": "R: Expose Prometheus metrics.\n\nReturns:\n    Prometheus text format metrics",
        "operationId": "metrics_metrics_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AskRes": {
        "properties": {
          "answer": {
            "type": "string",
            "title": "Answer"
          },
          "sources": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Sources"
          }
        },
        "type": "object",
        "required": [
          "answer",
          "sources"
        ],
        "title": "AskRes"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IngestBatchReq": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/IngestTextReq"
            },
            "type": "array",
            "maxItems": 10,
            "minItems": 1,
            "title": "Documents",
            "description": "List of documents to ingest (1-10)"
          }
        },
        "type": "object",
        "required": [
          "documents"
        ],
        "title": "IngestBatchReq"
      },
      "IngestBatchRes": {
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/IngestTextRes"
            },
            "type": "array",
            "title": "Documents"
          },
          "total_chunks": {
            "type": "integer",
            "title": "Total Chunks"
          }
        },
        "type": "object",
        "required": [
          "documents",
          "total_chunks"
        ],
        "title": "IngestBatchRes"
      },
      "IngestTextReq": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "description": "Document title (1-200 chars)"
          },
          "text": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 1,
            "title": "Text",
            "description": "Full document text to be chunked (1-100,000 chars)"
          },
          "source": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Optional source URL or identifier (max 500 chars)"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata",
            "description": "Additional custom metadata (JSONB)"
          }
        },
        "type": "object",
        "required": [
          "title",
          "text"
        ],
        "title": "IngestTextReq"
      },
      "IngestTextRes": {
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "chunks": {
            "type": "integer",
            "title": "Chunks"
          }
        },
        "type": "object",
        "required": [
          "document_id",
          "chunks"
        ],
        "title": "IngestTextRes"
      },
      "Match": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "format": "uuid",
            "title": "Chunk Id"
          },
          "document_id": {
            "type": "string",
            "format": "uuid",
            "title": "Document Id"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "score": {
            "type": "number",
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "document_id",
          "content",
          "score"
        ],
        "title": "Match"
      },
      "QueryReq": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Query",
            "description": "User's natural language question (1-2,000 chars)"
          },
          "top_k": {
            "type": "integer",
            "maximum": 20.0,
            "minimum": 1.0,
            "title": "Top K",
            "description": "Number of similar chunks to retrieve (1-20)",
            "default": 5
          }
        },
        "type": "object",
        "required": [
          "query"
        ],
        "title": "QueryReq"
      },
      "QueryRes": {
        "properties": {
          "matches": {
            "items": {
              "$ref": "#/components/schemas/Match"
            },
            "type": "array",
            "title": "Matches"
          }
        },
        "type": "object",
        "required": [
          "matches"
        ],
        "title": "QueryRes"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key with appropriate scope (ingest, ask, metrics)"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}